Module Reddit_api_async.Api

module Parameters : sig ... end

module Sequencer : sig ... end

A Sequencer.t represents an endpoint that cannot have multiple requests in flight simultaneously.

module Request : sig ... end

module Api_error : sig ... end

An Api_error.t represents a "normal" error when making an API request.

type 'a t = {

request : Request.t;

handle_response : (Cohttp.Response.t * Cohttp.Body.t) -> ('a, Api_error.t) Core_kernel.Result.t;

sequencer : Sequencer.t option;

}

A t represents the combinaton of an HTTP request to Reddit and a function for turning the HTTP response into a typed representation.

val map : 'a t -> f:('a -> 'b) -> 'b t

type 'a with_param_override := ?⁠param_list_override:((string * string list) list -> (string * string list) list) -> unit -> 'a

type 'a with_listing_params := ?⁠pagination:Reddit_api_kernel.Listing.Pagination.t -> ?⁠count:int -> ?⁠limit:int -> ?⁠show_all:unit -> 'a

A value of type _ with_listing_params is a function with optional arguments representing Reddit's "listing" pagination protocol.

see https://www.reddit.com/dev/api#listings

Reddit's listing docs

Endpoints

Account

val me : Reddit_api_kernel.Thing.User.t t with_param_override

val karma : Reddit_api_kernel.Karma_list.t t with_param_override

val trophies : Reddit_api_kernel.Thing.Award.t list t with_param_override

val friends : Reddit_api_kernel.User_list.t t with_param_override with_listing_params

val blocked : Reddit_api_kernel.User_list.t t with_param_override with_listing_params

val messaging : Reddit_api_kernel.User_list.t t with_param_override with_listing_params

val trusted : Reddit_api_kernel.User_list.t t with_param_override with_listing_params

Flair

val select_flair : ?⁠background_color:Parameters.Color.t -> ?⁠css_class:string -> ?⁠flair_template_id:Uuid.t -> ?⁠text:string -> ?⁠text_color:Parameters.Color.t -> subreddit:Reddit_api_kernel.Subreddit_name.t -> target:Parameters.Flair_target.t -> unit t with_param_override

Links and comments

val add_comment : ?⁠return_rtjson:bool -> ?⁠richtext_json:Reddit_api_kernel.Json.t -> parent:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t | `Message of Reddit_api_kernel.Thing.Message.Id.t ] -> text:string -> Reddit_api_kernel.Thing.Comment.t t with_param_override

val delete : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val edit : ?⁠return_rtjson:bool -> ?⁠richtext_json:Reddit_api_kernel.Json.t -> id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> text:string -> [> `Link of Reddit_api_kernel.Thing.Link.t | `Comment of Reddit_api_kernel.Thing.Comment.t ] t with_param_override

val hide : links:Reddit_api_kernel.Thing.Link.Id.t list -> unit t with_param_override

val unhide : links:Reddit_api_kernel.Thing.Link.Id.t list -> unit t with_param_override

val lock : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val unlock : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val mark_nsfw : link:Reddit_api_kernel.Thing.Link.Id.t -> unit t with_param_override

val unmark_nsfw : link:Reddit_api_kernel.Thing.Link.Id.t -> unit t with_param_override

val more_children : ?⁠limit_children:bool -> link:Reddit_api_kernel.Thing.Link.Id.t -> more_comments:Reddit_api_kernel.Thing.More_comments.Details.By_children.t -> sort:Parameters.Comment_sort.t -> [ `Comment of Reddit_api_kernel.Thing.Comment.t | `More_comments of Reddit_api_kernel.Thing.More_comments.t ] list t with_param_override

val report : ?⁠from_modmail:bool -> ?⁠from_help_desk:bool -> ?⁠additional_info:string -> ?⁠custom_text:string -> ?⁠other_reason:string -> ?⁠rule_reason:string -> ?⁠site_reason:string -> ?⁠sr_name:string -> target:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t | `Message of Reddit_api_kernel.Thing.Message.Id.t | `Modmail_conversation of Reddit_api_kernel.Thing.Modmail_conversation.Id.t ] -> reason:string -> unit t with_param_override

val report_award : award_id:string -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override

val save : ?⁠category:string -> id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val unsave : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val saved_categories : (Cohttp.Response.t * Cohttp.Body.t) t with_param_override

val send_replies : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> enabled:bool -> unit t with_param_override

val set_contest_mode : link:Reddit_api_kernel.Thing.Link.Id.t -> enabled:bool -> unit t with_param_override

val set_subreddit_sticky : ?⁠to_profile:bool -> link:Reddit_api_kernel.Thing.Link.Id.t -> sticky_state:Parameters.Sticky_state.t -> unit t with_param_override

val set_suggested_sort : link:Reddit_api_kernel.Thing.Link.Id.t -> sort:Parameters.Comment_sort.t option -> unit t with_param_override

val spoiler : link:Reddit_api_kernel.Thing.Link.Id.t -> unit t with_param_override

val unspoiler : link:Reddit_api_kernel.Thing.Link.Id.t -> unit t with_param_override

val store_visits : links:Reddit_api_kernel.Thing.Link.Id.t list -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override

val submit : ?⁠ad:bool -> ?⁠nsfw:bool -> ?⁠resubmit:bool -> ?⁠sendreplies:bool -> ?⁠spoiler:bool -> ?⁠flair_id:string -> ?⁠flair_text:string -> ?⁠collection_id:string -> ?⁠event_start:Core_kernel.Time.t -> ?⁠event_end:Core_kernel.Time.t -> ?⁠event_tz:string -> subreddit:Reddit_api_kernel.Subreddit_name.t -> title:string -> kind:Parameters.Link_kind.t -> (Reddit_api_kernel.Thing.Link.Id.t * Uri.t) t with_param_override

val vote : ?⁠rank:int -> direction:Parameters.Vote_direction.t -> target:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val info : Parameters.Info_query.t -> [ `Comment of Reddit_api_kernel.Thing.Comment.t | `Link of Reddit_api_kernel.Thing.Link.t | `Subreddit of Reddit_api_kernel.Thing.Subreddit.t ] list t with_param_override

Listings

val best : (?⁠include_categories:bool -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val links_by_id : links:Reddit_api_kernel.Thing.Link.Id.t list -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override

val comments : ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> ?⁠comment:Reddit_api_kernel.Thing.Comment.Id.t -> ?⁠context:int -> ?⁠depth:int -> ?⁠limit:int -> ?⁠showedits:bool -> ?⁠showmore:bool -> ?⁠sort:Parameters.Comment_sort.t -> ?⁠threaded:bool -> ?⁠truncate:int -> link:Reddit_api_kernel.Thing.Link.Id.t -> Reddit_api_kernel.Comment_response.t t with_param_override

val duplicates : (?⁠crossposts_only:bool -> ?⁠sort:Parameters.Duplicate_sort.t -> link:Reddit_api_kernel.Thing.Link.Id.t -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val hot : (?⁠location:string -> ?⁠include_categories:bool -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val new_ : (?⁠include_categories:bool -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val rising : (?⁠include_categories:bool -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val top : (?⁠since:Parameters.Historical_span.t -> ?⁠include_categories:bool -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val controversial : (?⁠since:Parameters.Historical_span.t -> ?⁠include_categories:bool -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val random : ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Link.Id.t t with_param_override

Private messages

val block_author : id:[< `Comment of Reddit_api_kernel.Thing.Comment.Id.t | `Message of Reddit_api_kernel.Thing.Message.Id.t ] -> unit t with_param_override

val collapse_message : messages:Reddit_api_kernel.Thing.Message.Id.t list -> unit t with_param_override

val uncollapse_message : messages:Reddit_api_kernel.Thing.Message.Id.t list -> unit t with_param_override

val compose_message : ?⁠g_recaptcha_response:string -> ?⁠from_subreddit:Reddit_api_kernel.Subreddit_name.t -> to_:Reddit_api_kernel.Username.t -> subject:string -> text:string -> unit t with_param_override

val delete_message : message:Reddit_api_kernel.Thing.Message.Id.t -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override

val read_message : messages:Reddit_api_kernel.Thing.Message.Id.t list -> unit t with_param_override

val unread_message : messages:Reddit_api_kernel.Thing.Message.Id.t list -> unit t with_param_override

val inbox : (?⁠include_categories:bool -> ?⁠mid:string -> mark_read:bool -> Reddit_api_kernel.Inbox_item.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val unread : (?⁠include_categories:bool -> ?⁠mid:string -> mark_read:bool -> Reddit_api_kernel.Inbox_item.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val sent : (?⁠include_categories:bool -> ?⁠mid:string -> Reddit_api_kernel.Thing.Message.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val comment_replies : (?⁠include_categories:bool -> ?⁠mid:string -> mark_read:bool -> Reddit_api_kernel.Inbox_item.Comment.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val subreddit_comments : (subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Comment.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

Moderation

val log : (?⁠mod_filter:Parameters.Mod_filter.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> ?⁠type_:string -> Reddit_api_kernel.Mod_action.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val reports : (?⁠location:string -> ?⁠only:Parameters.Links_or_comments.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> [ `Link of Reddit_api_kernel.Thing.Link.t | `Comment of Reddit_api_kernel.Thing.Comment.t ] Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val spam : (?⁠location:string -> ?⁠only:Parameters.Links_or_comments.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> [ `Link of Reddit_api_kernel.Thing.Link.t | `Comment of Reddit_api_kernel.Thing.Comment.t ] Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val modqueue : (?⁠location:string -> ?⁠only:Parameters.Links_or_comments.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> [ `Link of Reddit_api_kernel.Thing.Link.t | `Comment of Reddit_api_kernel.Thing.Comment.t ] Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val unmoderated : (?⁠location:string -> ?⁠only:Parameters.Links_or_comments.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> [ `Link of Reddit_api_kernel.Thing.Link.t | `Comment of Reddit_api_kernel.Thing.Comment.t ] Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val edited : (?⁠location:string -> ?⁠only:Parameters.Links_or_comments.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> [ `Link of Reddit_api_kernel.Thing.Link.t | `Comment of Reddit_api_kernel.Thing.Comment.t ] Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val accept_moderator_invite : subreddit:Reddit_api_kernel.Subreddit_name.t -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override

val approve : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val remove : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> spam:bool -> unit t with_param_override

val distinguish : ?⁠sticky:bool -> id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> how:Parameters.How_to_distinguish.t -> [> `Link of Reddit_api_kernel.Thing.Link.t | `Comment of Reddit_api_kernel.Thing.Comment.t ] t with_param_override

val ignore_reports : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val unignore_reports : id:[< `Link of Reddit_api_kernel.Thing.Link.Id.t | `Comment of Reddit_api_kernel.Thing.Comment.Id.t ] -> unit t with_param_override

val leavecontributor : subreddit:Reddit_api_kernel.Thing.Subreddit.Id.t -> unit t with_param_override

val leavemoderator : subreddit:Reddit_api_kernel.Thing.Subreddit.Id.t -> unit t with_param_override

val mute_message_author : message:Reddit_api_kernel.Thing.Message.Id.t -> unit t with_param_override

val unmute_message_author : message:Reddit_api_kernel.Thing.Message.Id.t -> unit t with_param_override

val stylesheet : subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Stylesheet.t t with_param_override

New modmail

val create_modmail_conversation : subject:string -> body:string -> subreddit:Reddit_api_kernel.Subreddit_name.t -> to_:Parameters.Modmail_recipient.t -> hide_author:bool -> Reddit_api_kernel.Modmail.Conversation.t t with_param_override

val search : (?⁠category:string -> ?⁠include_facets:bool -> ?⁠restrict_to_subreddit:Reddit_api_kernel.Subreddit_name.t -> ?⁠since:Parameters.Historical_span.t -> ?⁠sort:Parameters.Search_sort.t -> ?⁠types:Parameters.Search_type.Set.t -> query:string -> (Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t option * [ `Subreddit of Reddit_api_kernel.Thing.Subreddit.t | `User of Reddit_api_kernel.Thing.User.t ] Reddit_api_kernel.Listing.t option) t with_param_override) with_listing_params

Subreddits

val banned : (?⁠include_categories:bool -> ?⁠user:Reddit_api_kernel.Username.t -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Relationship.Ban.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val muted : (?⁠include_categories:bool -> ?⁠user:Reddit_api_kernel.Username.t -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Relationship.Mute.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val wiki_banned : (?⁠include_categories:bool -> ?⁠user:Reddit_api_kernel.Username.t -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Relationship.Ban.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val contributors : (?⁠include_categories:bool -> ?⁠user:Reddit_api_kernel.Username.t -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Relationship.Contributor.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val wiki_contributors : (?⁠include_categories:bool -> ?⁠user:Reddit_api_kernel.Username.t -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Relationship.Contributor.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val moderators : (?⁠include_categories:bool -> ?⁠user:Reddit_api_kernel.Username.t -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Relationship.Moderator.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val delete_subreddit_image : subreddit:Reddit_api_kernel.Subreddit_name.t -> image:Parameters.Subreddit_image.t -> unit t with_param_override

val search_subreddits_by_name : ?⁠exact:bool -> ?⁠include_over_18:bool -> ?⁠include_unadvertisable:bool -> query:string -> Reddit_api_kernel.Subreddit_name.t list t with_param_override

val create_or_edit_subreddit : ?⁠comment_score_hide_mins:int -> ?⁠wiki_edit_age:int -> ?⁠wiki_edit_karma:int -> all_original_content:bool -> allow_discovery:bool -> allow_images:bool -> allow_post_crossposts:bool -> allow_top:bool -> allow_videos:bool -> api_type:(string * string list) list -> collapse_deleted_comments:bool -> crowd_control_mode:bool -> description:string -> disable_contributor_requests:bool -> exclude_banned_modqueue:bool -> free_form_reports:bool -> g_recaptcha_response:string option -> header_title:string -> hide_ads:bool -> key_color:string -> lang:string -> link_type:Parameters.Link_type.t -> name:string -> original_content_tag_enabled:bool -> over_18:bool -> public_description:string -> restrict_commenting:bool -> restrict_posting:bool -> show_media:bool -> show_media_preview:bool -> spam_comments:Parameters.Spam_level.t -> spam_links:Parameters.Spam_level.t -> spam_selfposts:Parameters.Spam_level.t -> spoilers_enabled:bool -> subreddit:Reddit_api_kernel.Subreddit_name.t -> submit_link_label:string -> submit_text:string -> submit_text_label:string -> suggested_comment_sort:Parameters.Comment_sort.t -> title:string -> type_:Parameters.Subreddit_type.t -> wiki_mode:Parameters.Wiki_mode.t -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override

val submit_text : subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Submit_text.t t with_param_override

val subreddit_autocomplete : ?⁠limit:int -> ?⁠include_categories:bool -> ?⁠include_over_18:bool -> ?⁠include_profiles:bool -> query:string -> Reddit_api_kernel.Thing.Subreddit.t Reddit_api_kernel.Listing.t t with_param_override

val set_subreddit_stylesheet : ?⁠reason:string -> subreddit:Reddit_api_kernel.Subreddit_name.t -> stylesheet_contents:string -> unit t with_param_override

val subscribe : ?⁠skip_initial_defaults:bool -> action:Parameters.Subscription_action.t -> subreddits:Parameters.Subscription_list.t -> unit t with_param_override

val search_users : (?⁠sort:Parameters.Relevance_or_activity.t -> query:string -> Reddit_api_kernel.Thing.User.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val about_subreddit : subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Subreddit.t t with_param_override

val subreddit_settings : ?⁠created:bool -> ?⁠location:string -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Subreddit_settings.t t with_param_override

val subreddit_rules : subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Subreddit_rules.t t with_param_override

val subreddit_traffic : subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Subreddit_traffic.t t with_param_override

val get_sticky : ?⁠number:int -> subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Thing.Link.Id.t t with_param_override

val get_subreddits : (?⁠include_categories:bool -> relationship:Parameters.Subreddit_relationship.t -> Reddit_api_kernel.Thing.Subreddit.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val search_subreddits_by_title_and_description : (?⁠show_users:bool -> ?⁠sort:Parameters.Relevance_or_activity.t -> query:string -> Reddit_api_kernel.Thing.Subreddit.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val list_subreddits : (?⁠include_categories:bool -> ?⁠show_users:bool -> sort:Parameters.Subreddit_listing_sort.t -> Reddit_api_kernel.Thing.Subreddit.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

Users

val about_user : username:Reddit_api_kernel.Username.t -> Reddit_api_kernel.Thing.User.t t with_param_override

val list_user_subreddits : (?⁠include_categories:bool -> sort:Parameters.User_subreddit_sort.t -> Reddit_api_kernel.Thing.Subreddit.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val add_relationship : relationship:Parameters.Relationship_spec.t -> username:Reddit_api_kernel.Username.t -> duration:Parameters.Relationship_spec.Duration.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> ?⁠note:string -> ?⁠ban_reason:string -> ?⁠ban_message:string -> ?⁠ban_context:string -> unit t with_param_override

val remove_relationship : relationship:Parameters.Relationship_spec.t -> username:Reddit_api_kernel.Username.t -> ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> unit t with_param_override

Wiki

val add_wiki_editor : page:Reddit_api_kernel.Wiki_page.Id.t -> user:Reddit_api_kernel.Username.t -> unit t with_param_override

val remove_wiki_editor : page:Reddit_api_kernel.Wiki_page.Id.t -> user:Reddit_api_kernel.Username.t -> unit t with_param_override

val edit_wiki_page : ?⁠previous:Reddit_api_kernel.Wiki_page.Revision.Id.t -> ?⁠reason:string -> content:string -> page:Reddit_api_kernel.Wiki_page.Id.t -> (unit, Reddit_api_kernel.Wiki_page.Edit_conflict.t) Core_kernel.Result.t t with_param_override

val toggle_wiki_revision_visibility : page:Reddit_api_kernel.Wiki_page.Id.t -> revision:Reddit_api_kernel.Wiki_page.Revision.Id.t -> [ `Became_hidden | `Became_visible ] t with_param_override

val revert_wiki_page : page:Reddit_api_kernel.Wiki_page.Id.t -> revision:Reddit_api_kernel.Wiki_page.Revision.Id.t -> unit t with_param_override

val wiki_discussions : (page:Reddit_api_kernel.Wiki_page.Id.t -> Reddit_api_kernel.Thing.Link.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val wiki_pages : ?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> string list t with_param_override

val subreddit_wiki_revisions : (?⁠subreddit:Reddit_api_kernel.Subreddit_name.t -> Reddit_api_kernel.Wiki_page.Revision.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val wiki_page_revisions : (page:Reddit_api_kernel.Wiki_page.Id.t -> Reddit_api_kernel.Wiki_page.Revision.t Reddit_api_kernel.Listing.t t with_param_override) with_listing_params

val wiki_permissions : page:Reddit_api_kernel.Wiki_page.Id.t -> Reddit_api_kernel.Wiki_page.Permissions.t t with_param_override

val set_wiki_permissions : page:Reddit_api_kernel.Wiki_page.Id.t -> listed:bool -> level:Reddit_api_kernel.Wiki_page.Permissions.Level.t -> Reddit_api_kernel.Wiki_page.Permissions.t t with_param_override

val wiki_page : ?⁠compare_revisions:(string option * string option) -> page:Reddit_api_kernel.Wiki_page.Id.t -> Reddit_api_kernel.Wiki_page.t t with_param_override