Module Credentials.Refresh_token

Refresh_token credentials correspond to Reddit's "web app" and "installed-app" app types.

see https://praw.readthedocs.io/en/stable/tutorials/refresh_token.html
PRAW's documentation on refresh tokens for advice on obtaining a refresh token, which is currently outside the scope of this project.

see https://datatracker.ietf.org/doc/html/rfc6749#section-6
The RFC 6749 section describing the corresponding access token request.

type t = {

client_id : string;

client_secret : string option;	(*
This field is present for web apps and absent for installed apps.
*)

refresh_token : string;

}

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t