Module Connection.Error

A request via Connection may result in two different HTTP requests: a request to the explicitly named API endpoint and, as necessary, a request to retrieve an OAuth2 access token. This type distinguishes errors associated with each of these requests.

type 'endpoint_error t =

| Access_token_request_error of Access_token_request_error.t

| Endpoint_error of 'endpoint_error

val sexp_of_t : ('endpoint_error -> Sexplib0.Sexp.t) -> 'endpoint_error t -> Sexplib0.Sexp.t